backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.7.1
pymongo==4.9.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
email-validator==2.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.9
